1.0 spec draft: Add support for mipmapped textures
All textures are now stored with levels of mipmapping down to 1x1.

The TXMD and TXPX segments are merged into a single TXTR segment that
contains all texture-related data.

The ENDS segment was removed, and it's instead mandated that the TXTR
segment appears as the last segment in the file.

The maximum number of textures is increased from 512 to 65536.